% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_argument.R
\name{add_argument}
\alias{add_argument}
\title{Add an argument to a list}
\usage{
add_argument(
  list,
  conn,
  variable,
  operator,
  values,
  manual = FALSE,
  statement = NULL
)
}
\arguments{
\item{list}{The list to which the argument will be added.}

\item{conn}{The connection object or database connection string.}

\item{variable}{The variable or column name to be used in the argument.}

\item{operator}{The operator to be used in the argument (e.g., "greater", "between", "equal", "less").}

\item{values}{The values to be used in the argument.}

\item{manual}{A logical value indicating whether the argument should be created automatically or manually.
If \code{FALSE}, the argument is created automatically using the provided parameters.
If \code{TRUE}, the \code{statement} parameter must be specified.}

\item{statement}{The manual argument select statement to be used. Required if \code{manual = TRUE}.}
}
\value{
The updated list with the added argument.
}
\description{
This function adds an argument to a list based on specified conditions.
}
