% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_argument_sequence.R
\name{get_argument_sequence}
\alias{get_argument_sequence}
\title{Get argument sequence based on argument relation}
\usage{
get_argument_sequence(arguments, argument_relation)
}
\arguments{
\item{arguments}{The list of arguments.}

\item{argument_relation}{The specified argument relation. If "and", the sequence will be 1:length(arguments).
If "or", the sequence will be rep(1, length(arguments)).
If a vector is provided, it should have the same length as the number of arguments.}
}
\value{
A numeric vector representing the sequence of arguments.
}
\description{
This function returns the sequence of arguments based on the specified argument relation.
The argument relation determines the logical relationship between the arguments (e.g., "and", "or").
}
\examples{
# Get argument sequence with "and" relation
arguments <- list("arg1", "arg2", "arg3")
argument_sequence <- get_argument_sequence(arguments, argument_relation = "and")
argument_sequence

# Get argument sequence with "or" relation
arguments <- list("arg1", "arg2", "arg3")
argument_sequence <- get_argument_sequence(arguments, argument_relation = "or")
argument_sequence

# Get argument sequence with custom relation
arguments <- list("arg1", "arg2", "arg3")
argument_relation <- c("or", "and", "or")
argument_sequence <- get_argument_sequence(arguments, argument_relation)
argument_sequence
}
