% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_query_results.R
\name{merge_query_results}
\alias{merge_query_results}
\title{Merge query results from multiple tables}
\usage{
merge_query_results(query_results)
}
\arguments{
\item{query_results}{A list of tables containing the query results.
The list should include the following tables: observation_table, between_table, within_table,
condition_table, dataset_table, study_table, publication_table, task_table.}
}
\value{
A merged data frame with the query results from all the tables.
}
\description{
This function merges the query results from multiple tables into a single data frame.
It performs left joins on the tables based on specified join keys.
}
\examples{
# Merge query results
query_results <- list(
  observation_table = data.frame(observation_id = 1:3, value = c(10, 20, 30)),
  between_table = data.frame(observation_id = 1:3, between_var = c("A", "B", "C")),
  within_table = data.frame(observation_id = 1:3, within_var = c("X", "Y", "Z"))
)
merged_data <- merge_query_results(query_results)
merged_data
}
