% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_only_matches.R
\name{return_only_matches}
\alias{return_only_matches}
\title{Return only matching rows between two data frames}
\usage{
return_only_matches(data1, data2, id = NULL, full = FALSE)
}
\arguments{
\item{data1}{The first data frame.}

\item{data2}{The second data frame.}

\item{id}{The column(s) to select from the joined data frames. By default,
it selects all common columns.}

\item{full}{Logical indicating whether to return the full joined data frames.
If TRUE, it returns all columns from both data frames.}
}
\value{
A data frame containing the matching rows based on the specified columns.
If id is specified, it returns a data frame with only the selected column(s).
If full is TRUE, it returns the full joined data frames.
}
\description{
This function returns the rows that have matching values in specified columns
between two data frames. It performs an inner join by default and returns the
matching rows based on the common columns between the data frames.
}
\examples{
# Return matching rows based on common columns
return_only_matches(data1, data2)

# Return matching rows and select specific column(s)
return_only_matches(data1, data2, id = "column_name")

# Return the full joined data frames
return_only_matches(data1, data2, full = TRUE)
}
