% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_proper_ids.R
\name{return_proper_ids}
\alias{return_proper_ids}
\title{Return proper IDs from argument matches}
\usage{
return_proper_ids(argument_matches, argument_sequence)
}
\arguments{
\item{argument_matches}{A list of connected ID lists for each argument.}

\item{argument_sequence}{The sequence of the arguments.}
}
\value{
A list of proper IDs for each table based on the argument matches
and sequence.
}
\description{
This function takes a list of connected ID lists and a sequence and returns
the joined version of the IDs. It determines the proper IDs based on the
argument sequence, ensuring that the IDs are appropriately filtered and
combined according to the argument relationship.
}
\examples{
# Return the proper IDs based on argument matches and sequence
return_proper_ids(argument_matches, argument_sequence)

}
