% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_table_is_queried.R
\name{which_table_is_queried}
\alias{which_table_is_queried}
\title{Identify the table queried in SQL argument}
\usage{
which_table_is_queried(sql_argument)
}
\arguments{
\item{sql_argument}{The SQL argument to analyze.}
}
\value{
The table name that is being queried.
}
\description{
This function takes an SQL argument and extracts the table name that is being
queried in the "FROM" clause. It uses regular expressions to locate the table
name after the "FROM" keyword.
}
\examples{
# Extract the table name from an SQL argument
which_table_is_queried(sql_argument)

}
