% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_id_query.R
\name{write_id_query}
\alias{write_id_query}
\title{Generate SQL query for filtering IDs in a table}
\usage{
write_id_query(table_name, variable_name, id_vector)
}
\arguments{
\item{table_name}{The name of the table to query.}

\item{variable_name}{The name of the variable/column to filter on.}

\item{id_vector}{A vector of IDs to match against the variable.}
}
\value{
The generated SQL query.
}
\description{
This function generates an SQL query to filter rows in a table based on a
vector of IDs. It constructs a query that selects all columns from the given
table where the specified variable name matches any of the IDs in the vector.
}
\examples{
# Generate an SQL query to filter IDs in a table
write_id_query("my_table", "id_column", c(1, 2, 3))

}
